refactor: split file explorer
